refactor: use abstract redis and keep full backward compatibility
