+++ tests/Test/Prometheus/Predis/GaugeWithPrefixTest.php
@@ -4,8 +4,7 @@ declare(strict_types=1);
 
 namespace Test\Prometheus\Predis;
 
-use Predis\Client;
-use Prometheus\Storage\Redis;
+use Prometheus\Storage\Predis;
 use Test\Prometheus\AbstractGaugeTest;
 
 /**
@@ -17,9 +16,7 @@ class GaugeWithPrefixTest extends AbstractGaugeTest
 {
     public function configureAdapter(): void
     {
-        $connection = new Client(['host' => REDIS_HOST, 'prefix' => 'prefix:']);
-
-        $this->adapter = Redis::fromExistingConnection($connection);
+        $this->adapter = new Predis(['host' => REDIS_HOST], ['prefix' => 'prefix:']);
         $this->adapter->wipeStorage();
     }
 }
